fixing code smells

+++ src/main/java/ru/r2cloud/lora/LoraObservationRequest.java
@@ -120,7 +120,7 @@ public class LoraObservationRequest {
 			dos.writeByte(syncword);
 			dos.writeByte(10); // power
 			dos.writeShort(preambleLength);
-			dos.writeByte((int) gain);
+			dos.writeByte(gain);
 			dos.writeByte(ldro);
 			if (useCrc) {
 				dos.writeByte(1);
